refactor remaining classes
