+++ dialogflow-cx/validate_form_sample.py
@@ -1,6 +1,12 @@
+import agent_delegator as agd
+import auth_delegator as ad
 import dialogflow_sample as ds
-import sample_base as sb
+import intent_delegator as idy
+import page_delegator as pd
+import start_flow_delegator as sfd
 import test_case_delegator as tcd
+import turn
+import webhook_delegator as wd
 from google.cloud.dialogflowcx import Form, Fulfillment, ResponseMessage
 
 
@@ -52,7 +58,7 @@ class ValidateFormSample(ds.DialogflowSample):
         agent_display_name=None,
     ):
         self.set_auth_delegator(
-            sb.AuthDelegator(
+            ad.AuthDelegator(
                 self,
                 project_id=project_id,
                 quota_project_id=quota_project_id,
@@ -60,31 +66,31 @@ class ValidateFormSample(ds.DialogflowSample):
             )
         )
         self.set_agent_delegator(
-            sb.AgentDelegator(self, display_name=agent_display_name)
+            agd.AgentDelegator(self, display_name=agent_display_name)
         )
-        self.webhook_delegator = sb.WebhookDelegator(
+        self.webhook_delegator = wd.WebhookDelegator(
             self, display_name=self._WEBHOOK_DISPLAY_NAME, uri=webhook_uri
         )
-        self.intent_delegator = sb.IntentDelegator(
+        self.intent_delegator = idy.IntentDelegator(
             self, display_name=self._INTENT_DISPLAY_NAME
         )
-        self.page_delegator = sb.FulfillmentPageDelegator(
+        self.page_delegator = pd.FulfillmentPageDelegator(
             self,
             display_name=self._PAGE_DISPLAY_NAME,
             entry_fulfillment_text=self._PAGE_ENTRY_FULFILLMENT_TEXT,
         )
-        self.start_flow_delegator = sb.StartFlowDelegator(self)
-        self.start_page_delegator = sb.StartPageDelegator(self)
+        self.start_flow_delegator = sfd.StartFlowDelegator(self)
+        self.start_page_delegator = pd.StartPageDelegator(self)
 
         self.test_case_delegators = {}
         for display_name, test_config in self.TEST_CASES.items():
-            turn_0 = sb.Turn(
+            turn_0 = turn.Turn(
                 test_config["input_text"][0],
                 test_config["expected_response_text"][0],
                 self.page_delegator,
                 self.intent_delegator,
             )
-            turn_1 = sb.Turn(
+            turn_1 = turn.Turn(
                 test_config["input_text"][1],
                 test_config["expected_response_text"][1],
                 self.start_page_delegator,
